필드 타입 변경

+++ grants/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class Grant(models.Model):
     member = models.ForeignKey(verbose_name="휴가 대상자", to="members.Member", on_delete=models.CASCADE)
-    stock = models.IntegerField(verbose_name="부여된 휴가 일 수")
+    stock = models.FloatField(verbose_name="부여된 휴가 일 수")
 
 
 class GrantLeave(models.Model):
